% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employees_attrition.R
\docType{data}
\name{employees_attrition}
\alias{employees_attrition}
\title{Employees attrition data set}
\format{
A data frame with 1470 observations and 35 variables.

\describe{
\item{xxx}{xxx.}
}
}
\usage{
employees_attrition
}
\description{
A data set with attrition information of several companies from IBM HR Analytics Employee Attrition & Performance.
}
\examples{

employees_attrition

}
\keyword{datasets}
