% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employees_for_ttest.R
\docType{data}
\name{employees_for_ttest}
\alias{employees_for_ttest}
\title{HR data set for t-student testing}
\format{
A data frame with 550 observations.

\describe{
\item{employeeID}{An integer variable with the employee ID.}
\item{gender}{A factor variable with the employee gender.}
\item{salary20}{A numerical variable showing the montly salary of the employee in 2020.}
\item{perfomance20}{A numerical variable representing the employee individual performance (between 0 and 100) in 2020.}
\item{salary21}{A numerical variable showing the montly salary of the employee in 2021.}
\item{perfomance21}{A numerical variable representing the employee individual performance (between 0 and 100) in 2021.}
}
}
\usage{
employees_for_ttest
}
\description{
Dummy HR data set to illustrate hypothesis testing through the t-student.
}
\examples{

employees_for_ttest

}
\keyword{datasets}
