% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employees_opinion_dataset.R
\docType{data}
\name{employees_opinion}
\alias{employees_opinion}
\title{Employees opinion data set}
\format{
A data frame with 149 observations.

\describe{
\item{commentID}{An integer variable with the opinion ID.}
\item{comment}{A character variable with the employee opinion about their organization.}
\item{assessment}{A factor variable indicating if the opinion is positive or negative.}
}
}
\usage{
employees_opinion
}
\description{
A dataset of positive and negative employees’ opinions on how they feel about the organization where they are working.
}
\examples{

employees_opinion

}
\keyword{datasets}
