% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employees_opinion_ext_dataset.R
\docType{data}
\name{employees_opinion_ext}
\alias{employees_opinion_ext}
\title{Employees opinion data set}
\format{
A data frame with 14,249 observations.

\describe{
\item{employeeID}{An integer variable with the employee ID.}
\item{industry}{factor variable with the industry in which the company belongs.}
\item{company}{A factor variable with the company name.}
\item{score}{An integer variable with the global company assessment by the employee.}
\item{likes}{A character variable with the positive employee opinion about their company.}
\item{dislikes}{A character variable with the negative employee opinion about their company.}
}
}
\usage{
employees_opinion_ext
}
\description{
A dataset of employees’ opinions on how they feel about the organization where they are working.
}
\examples{

employees_opinion_ext

}
\keyword{datasets}
