% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifa_players.R
\docType{data}
\name{fifa_players}
\alias{fifa_players}
\title{FIFA PLAYERS data set}
\format{
A data frame with 18,207 observations.

\describe{
\item{ID}{An integer variable with an internal player ID.}
\item{Name}{A character variable with the name of the player.}
\item{LS, ST, RS, LW, LF, CF, RF, RW, LAM, CAM, RAM, LM, LCM, CM, RCM, RM, LWB, LDM, CDM, RDM, RWB, LB, LCB, CB, RCB, RB}{A character variable with the player core score on this position, and a margin depending on a random factor.}
\item{The rest of the variables} {The rest of the variables contains what the varible name says}
}
}
\usage{
fifa_players
}
\description{
A FIFA players data set with information about their characteristics and the teams where they are playing.
}
\examples{

fifa_players

}
\keyword{datasets}
