% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_dataset_2013.R
\docType{data}
\name{hr_dataset_2013}
\alias{hr_dataset_2013}
\title{HR data set to carry out an exploratory data analysis}
\format{

}
\usage{
hr_dataset_2013
}
\description{
Fake HR data set to illustrate an exploratory data analyses
}
\examples{

hr_dataset_2013

}
\keyword{datasets}
