% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm_employee_attrition_performance.R
\docType{data}
\name{ibm_employees_attrition_performance}
\alias{ibm_employees_attrition_performance}
\title{IBM Employees Attrition and Performance data set}
\format{
A data frame with 1472 observations and 35 variables.

\describe{
\item{Age}{Numerical Value - Employee Age.}
\item{Attrition}{Factor Value - If the Employee has Left the Company (0=no, 1=yes).}
\item{BusinessTravel}{Factor Value - Frequency of Travels (1=No Travel, 2=Travel Frequently, 3=Tavel Rarely).}
\item{DailyRate}{Numerical Value - Salary Level.}
\item{Department}{Factor Value - Employee Department where he/she works (1=HR, 2=R&D, 3=Sales).}
\item{DistanceFromHome}{Numerical Value - Distance between home and workeplace.}
\item{Education}{Numerical Value - Level of Education.}
\item{EducationField}{Factor Value - Field of Knowledge related to the Employee's Education (1=HR, 2=Life Sciences, 3=Marketing, 4=Medical Sciences, 5=Others, 6= Technical).}
\item{EmployeeCount}{Numerical Value - Emplooyee Count Number}
\item{EmployeeNumber}{Numerical Value - Employee ID Number}
\item{EnvironmentSatisfaction}{Numerical Value - Level of Environment Satisfaction.}
\item{Gender}{Factor Value - Employee Gender (1=Female, 2=Male).}
\item{HourlyRate}{Numerical Value - Hourly Salary.}
\item{JobInvolvement}{Numerical Value - Level of the Job Involvement.}
\item{JobLevel}{Numerical Value - Level of Job.}
\item{JobRole}{Factor Value - Employee Role (1=HC Rep, 2=HR, 3=Lab Technician, 4=Manager, 5=Managing Director, 6=Reasearch Director, 7=Research Scientist, 8=Sales Executieve, 9=Sales Representative).}
\item{JobSatisfaction}{Numerical Value - Level of Job Satisfaction.}
\item{MaritalStatus}{Factor Value - Marital Status (1=Divorced, 2=Married, 3=Single).}
\item{MonthlyIncome}{Numerical Value - Monthly Income by the Employee Work.}
\item{MonthlyRate}{Numerical Value - Monthly rate by the Employee Work.}
\item{NumCompaniesWorked}{Numerical Value - No. of Companies Worked At.}
\item{Over18}{Factor Value - If the Age is oiver 18 (1=Yes, 2=No).}
\item{OverTime}{Factor Value - If the Employee Spends more Time than Expected at Work (1=Yes, 2=No).}
\item{PercentSalaryHike}{Numerical Value - Percentage Increase in Salary.}
\item{PerformanceRating}{Numerical Value - Performance Rating.}
\item{RelationshipSatisfaction}{Numerical Value - Level of Relations Satisfaction.}
\item{StandardHours}{Numerical Value - Number of Standard Hours according to their Agreeement.}
\item{StockOptionLevel}{Numerical Value - Stock Options Level.}
\item{TotalWorkingYears}{Numerical Value - Total Years Worked in the Empployee Life.}
\item{TrainingTimesLastYear}{Numerical Value - Hours Spent in Training Last Year.}
\item{WorkLifeBalance}{Numerical Value - Perception about the Balance between Professional and Personal Life.}
\item{YearsAtCompany}{Numerical Value - Total Number of Years at the Company}
\item{YearsInCurrentRole}{Numerical Value - Years in the Current Role.}
\item{YearsSinceLastPromotion}{Numerical Value - Years from the Last Promotion.}
\item{YearsWithCurrManager}{Numerical Value - Years Spent With their Current Manager or Supervisor.}
}
}
\source{
\url{https://www.kaggle.com/pavansubhasht/ibm-hr-analytics-attrition-dataset}
}
\usage{
ibm_employees_attrition_performance
}
\description{
A data set with attrition and performance information of several companies from IBM HR Analytics Employee Attrition & Performance.
}
\examples{

ibm_employees_attrition_performance

}
\keyword{datasets}
